requirements.txt
fastapi[all]
psycopg2
websocket-client